add seed data, add components, reduce useStates and useEffects
