+++ src/frontend/src/pages/GanttPage/GanttChart.tsx
@@ -1,7 +1,6 @@
-import { Box, Chip, IconButton, Typography, useTheme } from '@mui/material';
+import { Box } from '@mui/material';
 import { EventChange, GanttTask, RequestEventChange } from '../../utils/gantt.utils';
-import { Edit } from '@mui/icons-material';
-import GanttChartSection from './GanttChartSection';
+import GanttChartTeamSection from './GanttChartTeamSection';
 
 interface GanttChartProps {
   startDate: Date;
@@ -28,8 +27,6 @@ const GanttChart = ({
   setShowWorkPackagesMap,
   highlightedChange
 }: GanttChartProps) => {
-  const theme = useTheme();
-
   return (
     <Box>
       {teamsList.map((teamName: string) => {
@@ -41,77 +38,22 @@ const GanttChart = ({
 
         const isEditMode = chartEditingState.find((entry) => entry.teamName === teamName)?.editing || false;
 
-        const handleEdit = () => {
-          const index = chartEditingState.findIndex((entry) => entry.teamName === teamName);
-          if (index !== -1) {
-            chartEditingState[index] = { teamName, editing: !isEditMode };
-          }
-
-          if (!isEditMode) {
-            const projects = tasks ? tasks.filter((event) => !event.project) : [];
-            projects.forEach((project) => {
-              setShowWorkPackagesMap((prev) => new Map(prev.set(project.id, true)));
-            });
-          }
-
-          setChartEditingState([...chartEditingState]);
-        };
-
-        if (!tasks) return <></>;
-
-        // Sorting the work packages of each project based on their start date
-        tasks.forEach((task) => {
-          task.children.sort((a, b) => a.start.getTime() - b.start.getTime());
-        });
-
-        return (
-          <Box
-            sx={{
-              mt: 1,
-              py: 1,
-              background: isEditMode ? theme.palette.divider : 'transparent',
-              borderRadius: '0.25rem',
-              width: 'fit-content'
-            }}
-          >
-            <Box
-              sx={{
-                display: 'flex',
-                alignItems: 'center',
-                gap: 2,
-                mb: '-15px',
-                pl: 2,
-                position: 'sticky',
-                left: 0,
-                width: 'fit-content',
-                height: '30px'
-              }}
-            >
-              <Typography variant="h6" fontWeight={400}>
-                {teamName}
-              </Typography>
-
-              {isEditMode ? (
-                <Chip label="Save" onClick={handleEdit} />
-              ) : (
-                <IconButton onClick={handleEdit}>
-                  <Edit />
-                </IconButton>
-              )}
-            </Box>
-            <Box key={teamName} sx={{ my: 0, width: 'fit-content', pl: 2 }}>
-              <GanttChartSection
-                tasks={tasks}
-                start={startDate}
-                end={endDate}
-                isEditMode={isEditMode}
-                saveChanges={saveChanges}
-                showWorkPackagesMap={showWorkPackagesMap}
-                setShowWorkPackagesMap={setShowWorkPackagesMap}
-                highlightedChange={highlightedChange}
-              />
-            </Box>
-          </Box>
+        return tasks ? (
+          <GanttChartTeamSection
+            startDate={startDate}
+            endDate={endDate}
+            chartEditingState={chartEditingState}
+            setChartEditingState={setChartEditingState}
+            saveChanges={saveChanges}
+            showWorkPackagesMap={showWorkPackagesMap}
+            setShowWorkPackagesMap={setShowWorkPackagesMap}
+            isEditMode={isEditMode}
+            teamName={teamName}
+            tasks={tasks}
+            highlightedChange={highlightedChange}
+          />
+        ) : (
+          <></>
         );
       })}
     </Box>
