+++ src/frontend/src/pages/GanttPage/GanttChartSection.tsx
@@ -4,17 +4,18 @@
  */
 
 import { eachDayOfInterval, isMonday } from 'date-fns';
-import { applyChangesToEvents, EventChange, GanttTaskData, RequestEventChange } from '../../utils/gantt.utils';
+import { EventChange, GanttTaskData, RequestEventChange } from '../../utils/gantt.utils';
 import { Box, Typography, Collapse } from '@mui/material';
-import GanttTaskBar from './GanttChartComponents/GanttTaskBar';
-import { useEffect, useState } from 'react';
+import GanttTaskBar from './GanttChartComponents/GanttTaskBar/GanttTaskBar';
+import { useState } from 'react';
+import GanttToolTip from './GanttChartComponents/GanttToolTip';
 
 interface GanttChartSectionProps {
   start: Date;
   end: Date;
-  tasks: GanttTaskData[];
+  displayEvents: GanttTaskData[];
   isEditMode: boolean;
-  saveChanges: (eventChanges: EventChange[]) => void;
+  createChange: (change: EventChange) => void;
   showWorkPackagesMap: Map<string, boolean>;
   setShowWorkPackagesMap: React.Dispatch<React.SetStateAction<Map<string, boolean>>>;
   highlightedChange?: RequestEventChange;
@@ -23,40 +24,38 @@ interface GanttChartSectionProps {
 const GanttChartSection = ({
   start,
   end,
-  tasks,
+  displayEvents,
   isEditMode,
-  saveChanges,
+  createChange,
   showWorkPackagesMap,
   setShowWorkPackagesMap,
   highlightedChange
 }: GanttChartSectionProps) => {
   const days = eachDayOfInterval({ start, end }).filter((day) => isMonday(day));
-  const [eventChanges, setEventChanges] = useState<EventChange[]>([]);
+  const projects = displayEvents.filter((event) => !event.project);
+  const [currentTask, setCurrentTask] = useState<GanttTaskData | undefined>(undefined);
+  const [cursorX, setCursorX] = useState(0);
+  const [cursorY, setCursorY] = useState(0);
 
-  const createChange = (change: EventChange) => {
-    setEventChanges([...eventChanges, change]);
+  const handleOnMouseOver = (e: React.MouseEvent, event: GanttTaskData) => {
+    if (!isEditMode) {
+      setCursorX(e.clientX);
+      setCursorY(e.clientY);
+      setCurrentTask(event);
+    }
   };
 
-  useEffect(() => {
-    // only try to save changes when we're going from non-editing to editing mode
-    if (!isEditMode) {
-      saveChanges(eventChanges);
-      setEventChanges([]); // reset the changes after sending them
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [isEditMode]);
-
-  const displayEvents = applyChangesToEvents(tasks, eventChanges);
-  const projects = displayEvents.filter((event) => !event.project);
+  const handleOnMouseLeave = () => {
+    setCurrentTask(undefined);
+  };
 
   const toggleWorkPackages = (projectTask: GanttTaskData) => {
     setShowWorkPackagesMap((prev) => new Map(prev.set(projectTask.id, !prev.get(projectTask.id))));
   };
 
-  return tasks.length > 0 ? (
+  return displayEvents.length > 0 ? (
     <Box sx={{ width: 'fit-content' }}>
-      {/* Data display: reset list of events every time eventChanges list changes using key */}
-      <Box sx={{ mt: '1rem', width: 'fit-content' }} key={eventChanges.length}>
+      <Box sx={{ mt: '1rem', width: 'fit-content' }}>
         {projects.map((project) => {
           return (
             <>
@@ -67,6 +66,8 @@ const GanttChartSection = ({
                   event={project}
                   isEditMode={isEditMode}
                   createChange={createChange}
+                  handleOnMouseOver={handleOnMouseOver}
+                  handleOnMouseLeave={handleOnMouseLeave}
                   onWorkPackageToggle={() => toggleWorkPackages(project)}
                   showWorkPackages={showWorkPackagesMap.get(project.id)}
                 />
@@ -81,6 +82,8 @@ const GanttChartSection = ({
                       event={displayWorkPackage!}
                       isEditMode={isEditMode}
                       createChange={createChange}
+                      handleOnMouseOver={handleOnMouseOver}
+                      handleOnMouseLeave={handleOnMouseLeave}
                       highlightedChange={
                         highlightedChange && workPackage.id === highlightedChange.eventId ? highlightedChange : undefined
                       }
@@ -92,6 +95,18 @@ const GanttChartSection = ({
           );
         })}
       </Box>
+      {currentTask && (
+        <GanttToolTip
+          xCoordinate={cursorX}
+          yCoordinate={cursorY}
+          title={!currentTask.project ? currentTask.name.substring(8) : currentTask.name.substring(6)}
+          startDate={currentTask.start}
+          endDate={currentTask.end}
+          color={currentTask.styles?.backgroundColor}
+          projectLead={currentTask.projectLead}
+          projectManager={currentTask.projectManager}
+        />
+      )}
     </Box>
   ) : (
     <Typography sx={{ mx: 1 }}>No items to display</Typography>
